# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateListPayload": {
        "description": "候选列表读取：完整详情、数量与全部入选日期。",
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidatePayload"
            },
            "title": "Candidates",
            "type": "array"
          },
          "dates": {
            "items": {
              "type": "string"
            },
            "title": "Dates",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/CandidateSummaryPayload"
          }
        },
        "required": [
          "candidates",
          "summary",
          "dates"
        ],
        "title": "CandidateListPayload",
        "type": "object"
      },
...
      "CandidateScope": {
        "description": "候选归类视图范围：待归类含未查看、已查看未决策与稍后处理，已处理含其余终态。",
        "enum": [
          "unprocessed",
          "processed"
        ],
        "title": "CandidateScope",
        "type": "string"
      },
...
      "CandidateState": {
        "description": "候选项处理状态：未处理（待归类/稍后处理）与三类已处理终态。",
        "enum": [
          "pending",
          "later",
          "dismissed",
          "observed",
          "cleared"
        ],
        "title": "CandidateState",
        "type": "string"
      },
...
      "FailurePayload": {
        "description": "业务失败响应：哪个状态码适用仍由各端点自己声明（`responses`）。",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/FailureDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "FailurePayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DailyScreen Lite",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/classification/candidates": {
      "get": {
        "operationId": "list_candidates_api_classification_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CandidateScope",
              "default": "unprocessed"
            }
          },
          {
            "in": "query",
            "name": "importDate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Importdate"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "result",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CandidateState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateListPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Candidates",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification/candidates/{candidate_id}": {
      "get": {
        "operationId": "get_candidate_api_classification_candidates__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "候选项不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification/candidates/{candidate_id}/dismiss": {
      "post": {
        "description": "暂不关注：只结束本次归类，不删除导入事实。",
        "operationId": "dismiss_api_classification_candidates__candidate_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateActionPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "命令无效或当前状态不允许"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "候选项不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismiss",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification/candidates/{candidate_id}/focus": {
      "post": {
        "description": "手动打开候选卡：作为浏览路径的新一步，不改筛选、不改未处理池顺序。\n\n目标不在当前筛选结果里也照常打开，响应里的 `inFilter` 让界面明确提示；\n已处理候选同样可以打开查看（自动推进仍只在待归类范围内找人）。",
        "operationId": "focus_api_classification_candidates__candidate_id__focus_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "候选项不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Focus",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification/candidates/{candidate_id}/later": {
      "post": {
        "description": "稍后处理：仍在待归类池，移到队尾。",
        "operationId": "later_api_classification_candidates__candidate_id__later_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateActionPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "命令无效或当前状态不允许"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "候选项不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Later",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification/candidates/{candidate_id}/reclassify": {
      "post": {
        "description": "主动重新归类：复用同一候选项，保留历史并调到队首。",
        "operationId": "reclassify_api_classification_candidates__candidate_id__reclassify_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateActionPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "命令无效或当前状态不允许"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "候选项不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reclassify",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification/cleanup": {
      "post": {
        "description": "主动清理待归类池；可限定入选日期。不删除导入事实。\n\n没有请求体或 `importDate` 为 null 都表示不限日期（同迁移前）。",
        "operationId": "cleanup_api_classification_cleanup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CleanupCommand"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cleanup",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification/search": {
      "get": {
        "description": "全局搜索：只返回经过导入与证券识别的股票，供跳转到候选归类。",
        "operationId": "search_api_classification_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateSearchPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "classification"
        ]
      }
    },
...
  }
}